+++ python/lsst/ts/mtrotator/__init__.py
@@ -19,14 +19,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+try:
+    from .version import *
+except ImportError:
+    __version__ = "?"
+
 from .constants import *
 from .enums import *
 from .structs import *
 from .mock_controller import *
 from .rotator_commander import *
 from .rotator_csc import *
-
-try:
-    from .version import *
-except ImportError:
-    __version__ = "?"